backend/requirements.txt
# LangGraph Agentic Workflow Dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0

# YouTube transcript extraction
youtube-transcript-api>=0.6.2

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
httpx>=0.27.0

# API server
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.9.0
mutagen>=1.47.0
elevenlabs>=1.3.0